Add Verification for installing snap programs

+++ src/software-snap.sh
@@ -1,20 +1,29 @@
 #!/bin/bash
 
 # Install app snap
-sudo apt-get -y install snapd
-sudo systemctl stop snapd.service
-sudo systemctl start snapd.service
+if [ -z $(which snap) ]
+then
+    sudo apt-get -y install snapd
+    sudo systemctl stop snapd.service
+    sudo systemctl start snapd.service
+fi
 
 sleep 3
 
 SOFTWARE_TO_INSTALL_SNAP_CLASSIC=$(cat ./softwares/software-snap-classic.txt)
 
 for software in ${SOFTWARE_TO_INSTALL_SNAP_CLASSIC[@]}; do
-    sudo snap install $software --classic
+    if [ -z $(which ${software}) ]
+    then
+        sudo snap install $software --classic
+    fi
 done
 
 SOFTWARE_TO_INSTALL_SNAP=$(cat ./softwares/software-snap.txt)
 
 for software in ${SOFTWARE_TO_INSTALL_SNAP[@]}; do
-    sudo snap install $software
+    if [ -z $(which ${software}) ]
+    then
+        sudo snap install $software
+    fi
 done
